+++ jirkov.vladimir/P4/inputArray.hpp
@@ -3,7 +3,7 @@
 #include <iosfwd>
 namespace Array
 {
-  size_t inputArray(std::istream &, int *matrix, size_t Size);
-  void printArray(std::ostream &, int const *matrix, size_t m, size_t n);
+  int inputArray(std::istream &, int *matrix, int Size);
+  void printArray(std::ostream &, int const *matrix, int m, int n);
 }
 #endif
